Add Troubleshoot menu to PublicScreens
